requirements/dev.txt
flake8==7.0.0
python-dotenv==1.0.1
Django==4.2
django-debug-toolbar==4.3.0
p8-naming==0.13.3
flake8-import-order==0.18.2
